Aws setup

+++ server/api/uploads.js
@@ -6,38 +6,66 @@ const path = require('path')
 const {
   models: { Image },
 } = require('../db/index');
+const AWS = require('aws-sdk')               
+const multer = require('multer')
 
-var multer = require('multer')
-const storage = multer.diskStorage({
-  destination: (req, file, cb) => {
-    cb(null, './server/images');
-  },
-  filename: async (req, file, cb, )=> {
-    console.log(req.query)
-    let {userId} = req.query;
-    const ext = path.extname(file.originalname)
-    const filepath = `/${uuidv4()}${ext}`
-      try {
-        const savedImage = await Image.create({
-          filepath: filepath,
-          userId: userId
-        });
-        req.imgData = savedImage
-        cb(null, filepath);
-      } catch (err){
-        console.log('Save Image Error:', err)
-      }
+const storage = multer.memoryStorage({
+    destination: function (req, file, cb) {
+        cb(null, '')
     }
-  }
-)
-var upload = multer({ storage });
+})
 
-router.post('/', upload.single('uploaded_file'), (req, res)=> {
-    res.json({
-      result:[req.imgData],
-    }).status(204).end()
-});
+// const filefilter = (req, file, cb) => {
+//     if (file.mimetype === 'image/jpeg' || file.mimetype === 'image/jpg') {
+//         cb(null, true)
+//     } else {
+//         cb(null, false)
+//     }
+// }
 
+const upload = multer({ storage: storage });
+
+const s3 = new AWS.S3({
+    accessKeyId: process.env.AWS_ACCESS_KEY_ID,             
+    secretAccessKey: process.env.AWS_ACCESS_KEY_SECRET,
+    region: 'us-east-2'
+})
+
+router.post('/', upload.single('uploaded_file'), async(req, res) => {
+    try{
+        const {userId} = req.query
+        const ext = path.extname(req.file.originalname)
+        const newFilepath = `${uuidv4()}${ext}`
+
+
+        const params = {
+            Bucket: process.env.S3_BUCKET_NAME,   
+            Key: newFilepath,           
+            Body:req.file.buffer,                   
+            ACL:"public-read-write",                 
+            ContentType:"image/jpeg"        
+        };
+
+         const uploadedImage = await s3.upload(params,(error,data)=>{
+            console.log('data',data)}).promise()
+
+          const savedImage = await Image.create({
+            filepath: uploadedImage.Location,
+            fileName: req.file.originalname,
+            userId: userId
+            });
+  
+          req.imgData = savedImage
+  
+          res.json({
+              result:[req.imgData],
+          }).status(204).end()
+
+    } catch(error){
+        console.log('Save Image Error:', error)
+        res.status(500).send({"err":error}) 
+    }
+})
 
 router.get('/', async (req,res)=>{
   try {
@@ -62,9 +90,10 @@ router.delete('/:id', async (req,res, next)=>{
       }
     })
     let filePath = "./server/images"+req.body.filePath
-    fs.unlink(filePath, (err) => {
-      if (err) throw err;
-      console.log('file was deleted')})
+    if(fs.existsSync(filePath)){
+      fs.unlink(filePath, (err) => {
+        if (err) throw err;
+        console.log('file was deleted')})}
     res.json(image)
 } catch (err) {
     console.log(err);
